% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_charts.R
\name{connected_scatter}
\alias{connected_scatter}
\title{Creates a connected_scatter plot}
\usage{
connected_scatter(df, plot_title = "Title", plot_caption = "", row_col, color_list = c())
}
\arguments{
\item{df}{Two-way percent distribution table}

\item{plot_title}{Plot title. Requires a string.}

\item{plot_caption}{Plot footnote. Requries a string.}

\item{row_col}{if "row", uses the rows as the group (each row will be a line); if "col" uses the column as the group (each column will correspond to a line)}

\item{color_list}{list of colors of lines}
}
\value{
connected_scatter bar chart
}
\description{
Create a connected_scatter plot from a two-way percent distribution table for up to three groupings of variables.
}
\examples{
# Load in the historical panel data
panel <- read_sheddata(historicalpanel)

#wellbeing by year
wellbeing <- tabrow(shed_data = panel, row_vars = c(year), col_vars = c(atleast_okay), colval1 = c(1),
labelcolvars = T, overall=F)

connected_scatter(wellbeing, plot_title = "Doing okay financially by year", row_col = "col")

# Figure 2 in SHED 2017 Report
tab <- c()
for(y in 2013:2017){
 tab_year <- tabrow(panel \%>\% filter(year == y),
                     row_var = educ_3cat,
                     col_var = atleast_okay,
                     rowval1 = c(1:3),
                     colval1 = 1,
                     overall = F)
  names(tab_year)[2] <- y

  if(y != 2013){
    tab_year <- tab_year[,2]
  }
  tab <- bind_cols(tab, tab_year)
  }
fig2 <- connected_scatter(tab,
                          plot_title = "Figure 2. At least doing okay financially (by survey year and education)",
                          row_col = "col")


}
