% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{freqtable}
\alias{freqtable}
\title{freqtable}
\usage{
#method for defualt
}
\arguments{
\item{shed_data}{SHED dataset.}

\item{freq_var}{the variable to create a tab of across other variables}

\item{freq_val}{the value of the frequency variable}

\item{col_var}{the column variable}

\item{overall}{if TRUE, includes overall row within each group.}

\item{weight}{weight variable used to create percent distribution, default is population weight (final_weight)}

\item{row_var}{the row variable}

\item{colname}{specifies the name of column [n], up to 5 columns}

\item{missing}{if TRUE, includes missing values in the denominator when calculating frequencies}
}
\value{
the output will be a table
}
\description{
This function takes in a SHED dataset and two lists of variables. It then reports the relative frequency
}
\examples{
fig28 <- freqtable(shed_2019,
                 col_var = degree_enrollment,
                  col_val = c(1,2,3),
                  row_var = ppethm,
                  rowvalall = c(1,2,4),
                  freq_var = benefits,
                  freq_val = c(1))

}
