% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_charts.R
\name{grouped_bar}
\alias{grouped_bar}
\title{Creates a grouped bar chart}
\usage{
grouped_bar(df, plot_title = "Title", plot_caption = "", row_or_col)
}
\arguments{
\item{df}{Two-way percent distribution table}

\item{plot_title}{Plot title. Requires a string.}

\item{plot_caption}{Plot footnote. Requries a string.}

\item{row_col_subrow}{if row, uses the row as the group; if col uses the column as the group (should typically use row if df from tabrow), if subrow recognizes format from tabsubrow}

\item{rev}{if True, colors go from light to dark if horizontal, if false colors go from dark to light if horiztonal, opposite if vertical}

\item{horizontal}{if True, returns a horizontal grouped bar chart; if False, returns a vertical grouped bar chart}

\item{bold_rows}{Adds bold labelling to each row number in list.

 @return Grouped bar chart}
}
\description{
Create a grouped chart from a two-way percent distribution table
}
\examples{
# Load in the SHED 2019 data
shed_2019 <- read_sheddata(2019)
# Figure 11 in SHED 2019 Report
fig3_4_data <- tabcol(shed_2019, row_vars = c(D34), col_vars = educ_3cat, rowvalall = c(2,3), labelrowvars = T, labelcolvars = F)
fig3_4_chart <- grouped_bar(fig3_4_data, row_col_subrow = "row", plot_title = "Figure 11. Usual place of work (by education)", plot_caption = "Among adults who worked for someone else")
}
