% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{multi1way}
\alias{multi1way}
\title{Configure table and charts for SHED annual report}
\usage{
#method for default
multi1way(shed_data, var_list, rowval1 = NA, rowval2 = NA, rowval3 = NA, rowval4 = NA, rowval5 = NA,rowval6 = NA, rowval7 = NA, rowval8 = NA,
rowval9 = NA, rowval10 = NA, rowval11 = NA, rowvalall=NA, breaks=NA,labelbreaks= FALSE, labelvarnames= TRUE,
colname1 = "Characteristic", weight = final_weight, missing = FALSE, breaklabels = NA, roundn = 0)
}
\arguments{
\item{shed_data}{SHED dataset}

\item{var_list}{a list of SHED variables for percent distribution table}

\item{rowvalall}{only displays rows where each SHED variable in var_list has a value in rowvalall}

\item{breaks}{Inserts a row break before each specified row variable.}

\item{auto_labelbreaks}{if TRUE, displays respective row variable label in row break lines. if FALSE, breaks are left as empty rows, NA.}

\item{labelvarnames}{if TRUE, displays row variable in row break lines. if FALSE, displays value labels in rows. Default is to show value labels in rows. Usually only specified
when using one value per variable (i.e. share saying yes) so the variable name is more meaningful than
the value label.}

\item{colname1}{specifies the first column name}

\item{weight}{weight variable used to create percent distribution, default is population weight (final_weight)}

\item{missing}{if TRUE, includes missing values in the denominator when calculating frequencies
if FALSE, excludes missing values in the denominator when calculating frequencies. If missing values are found in row
variables, this options must be specified.}

\item{breaklabels}{specifies a list of strings for labeling the breaks, where the first object in the string corresponds to the first break}

\item{roundn}{integer indicating the number of decimal places (round).}

\item{rowval}{for rowval[n], only display rows where the nth row variable has a value in numlist. If not specified, the default is all values.}
}
\value{
the output will be a dataframe of a percent distribution table.
dev
}
\description{
multi1way () is similar to the stata tab command but with flexibility to include multiple variables to the table
option to add breaks and to add labels to the breaks. It allows for additional options such as including all adults or to just
respondents the question was asked to.
}
\examples{
# Creates a dataframe using the SHED 2019 dataset with two break labels and label breaks before the first and third specified variables.
data("shed2019")
multi1way(shed2019, c(R1_f, R1_e, R1_a, R1_c, R1_b, R1_d), weight = final_weight, missing = FALSE, labelvarnames = TRUE,
                 rowvalall = 1, breaks =c(1,3), breaklabels=c("Mortgage Access", "Preference"))

# Creates family income distribution table
fig2.4_data <- multi1way(shed_2019, I40, rowvalall = c(0:11), colname1 = "Income Category")
}
