% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{one_var_tab}
\alias{one_var_tab}
\title{Create one way table for SHED variable}
\usage{
#default option
one_var_tab <- function(shed_data, shed_var, weight = final_weight, missing = FALSE, roundn = 0)
}
\arguments{
\item{shed_data}{SHED dataset}

\item{shed_var}{a SHED variable}

\item{weight}{weight variable used to create percent distribution, default is population weight (final_weight)}

\item{missing}{if TRUE, includes missing values in the denominator when calculating frequenices.
FALSE: excludes missing values in the denominator when calculating frequencies. If missing values are found in the row variables,
this option must be specified}

\item{roundn}{integer indicating the number of decimal places (round).}
}
\value{
the output will be a dataframe that is a percent distribution table
}
\description{
one_var_tab() creates percent distribution table for a single variable
}
\examples{
data("shed2019")
one_var_tab(shed2019, B2, weight=TRUE, missing=TRUE)
}
