% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_charts.R
\name{pie_chart}
\alias{pie_chart}
\title{Creates a simple pie chart}
\usage{
pie_chart(df, plot_title = "Title", plot_caption = "")
}
\arguments{
\item{df}{One-way percent distribution table}

\item{plot_title}{Plot title. Requires a string.}

\item{plot_caption}{Plot footnote. Requries a string.}
}
\value{
Pie chart
}
\description{
Create a simple pie chart from a one-way percent distribution table
}
\examples{
# Load in the SHED 2019 data
shed_2019 <- read_sheddata(2019)
bank_status_table <- one_var_tab(shed_2019, bank_status)
# Create bar chart from the one-way table
pie_chart(bank_status_table, plot_title = "Figure 18. Banking Status", plot_caption = "Fully banked individuals had a bank or credit union account \\n and had not used an alternative financial service in the past year.")
}
