% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_sheddata}
\alias{read_sheddata}
\title{Read internal SHED data}
\usage{
read_sheddata(year)
}
\arguments{
\item{Typically}{the input for this function is the year (ranging from 2013-2019). However for historical and shed supplement datasets, the input will be the historical_panel, shedapril2020 or shedjuly2020.}
}
\value{
The output will be a tibble.
}
\description{
This function reads in Federal Reserve SHED data for each available year. The function also serves to read in historical and public use SHED datasets
This dataset is intended only for the internal use of the Consumer and Community Research team.
To load in the historical dataset, use year = "historicalpanel". To load in the COVID supplement data,
use year = "april2020" or year = "july2020"
}
\examples{
# Create a dataframe for SHED 2017.
df <- read_sheddata(2017)
# Create a dataframe for historical SHED dataset
df <- read_sheddata(historical_panel)
}
