% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shed2014}
\alias{shed2014}
\title{Survey of Household Economics and Decisionmaking (SHED) 2014 Data}
\format{A tibble with 5896 observations and 320 variables}
\source{
<https://www.federalreserve.gov/consumerscommunities/shed_data.htm>
}
\usage{
data(shed2014)
}
\description{
Data from the 2014 Federal Reserve's Survey of Household Economic Decision Making (SHED).
This dataset measures the economic well-being of U.S. households and identifies potential risks to their finances.
The survey includes modules on a range of topics relevant to financial well-being including credit access and behaviors
savings, retirement, economic fragility, and education and student loans. In addition, the survey includes demographic
information about the respondent such as income, race/ethnicity, gender, education and more.
}
\details{
Description of each variable can be found in the codebook online, at <https://www.federalreserve.gov/consumerscommunities/files/shed_2014codebook.pdf>.

The variable final_weight corresponds to the final weight variable used to match characteristics of the U.S. adult population. For 2014,
final_weight is equal to weight3, which is scaled to match the number of respondents in the survey (sample weight). Note that population weights
are not provided in surveys before 2016.
}
\examples{
 data(shed2014)
 # load package for analyzing survey data and
 # calculate weighted tabulation of financial wellbeing
 library(questionr)
 wtd.table(x = shed2014$B2, weights = shed2014$final_weight)

}
\keyword{datasets}
