% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shedjuly2020}
\alias{shedjuly2020}
\title{Survey of Household Economics and Decisionmaking (SHED) July 2020 Supplement Data}
\format{A tibble with 4147 observations and 74 variables}
\source{
<https://www.federalreserve.gov/consumerscommunities/files/SHED-2019-codebook-supplemental-survey-july-2020.pdf>
}
\usage{
data(shedjuly2020)
}
\description{
Data from the July 2020 Federal Reserve's Survey of Household Economic Decision Making (SHED) Supplement.
This dataset measures the economic well-being of U.S. households and identifies potential risks to their finances
through the COVID-19 crisis.
}
\details{
Description of each variable can be found in the codebook online, at <https://www.federalreserve.gov/consumerscommunities/files/SHED-2019-codebook-supplemental-survey-july-2020.pdf>.

The variable final_weight corresponds to the final weight variable used to match characteristics of the U.S. adult population. For 2013,
final_weight is equal to weight, which is scaled to match the number of respondents in the survey (sample weight). Note that population weights
are not provided in surveys before 2016.
}
\examples{
 data(shed_july2020)
 # load package for analyzing survey data and
 # calculate weighted tabulation of financial wellbeing
 library(questionr)
 wtd.table(x = shedjuly2020$B2, weights = shedjuly2020$final_weight)

}
\keyword{datasets}
