% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_charts.R
\name{stacked_bar}
\alias{stacked_bar}
\title{Creates a stacked bar chart}
\usage{
stacked_chart(df, plot_title = "Title", plot_caption = "", row_col_subrow)
}
\arguments{
\item{df}{Two-way percent distribution table}

\item{plot_title}{Plot title. Requires a string.}

\item{plot_caption}{Plot footnote. Requries a string.}

\item{row_col_subrow}{if row, uses the row as the group; if col uses the column as the group; if subrow uses subrow as group (should typically use col if df from tabcol)}

\item{bold_rows}{Adds bold labelling to each row number in list.}
}
\value{
Stacked bar chart
}
\description{
Create a stacked chart from a two-way percent distribution table
}
\examples{
# Load in the SHED 2019 data
shed_2019 <- read_sheddata(2019)
# Figure 26 in SHED 2019 report
fig6_6_data <- tabcol(shed_2019, row_vars = c(GH1), col_vars = ppmsacat, rowvalall = c(2, 1, 3, 4), colvalall = c(1,2), labelrowvars = T, labelcolvars = F, overall  = F)
fig6_6_chart <- stacked_bar(fig6_6_data, "Housing tenure (by urban/rural status)", "Key identifies bars in order from left to right.", row_col_subrow = "col")

}
