% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tabrow}
\alias{tabrow}
\title{Title}
\usage{
#method for default
tabrow <- function(shed_data, row_vars, col_vars, rowval1 = NA, rowval2 = NA, rowval3 = NA, rowval4 = NA, rowval5 = NA,
rowval6 = NA, rowval7 = NA, rowval8 = NA, rowval9 = NA, rowval10 = NA, rowval11 = NA, rowval12 = NA, rowval13 = NA, rowvalall=NA,
colval1 = NA, colval2 = NA, colval3 = NA, colval4 = NA, colval5 = NA, colval6 = NA, colval7 = NA, colval8 = NA, colval9 = NA,
colval10 = NA, colval11 = NA, colval12 = NA, colval13 = NA, colvalall=NA, breaks=c(), auto_labelbreaks= FALSE, labelcolvars= FALSE, labelrowvars = FALSE,
colname1 = "Characteristic", weight = final_weight, missing = FALSE, breaklabels =c(), overall = TRUE, roundn = 0)
}
\arguments{
\item{shed_data}{SHED dataset}

\item{row_vars}{list of SHED variables used for the rows of the table}

\item{col_vars}{list of SHED variables used for the columns of the table}

\item{rowvalall}{for all row varibles, only displays rows where each SHED variable in var_list has a value in rowvalall}

\item{colvalall}{for all column variables, only displays columns where column variables have a value in colvalall}

\item{breaks}{inserts a row break before each new row}

\item{auto_labelbreaks}{if TRUE, displays row variable in row break lines. if FALSE, displays value labels in rows}

\item{labelcolvars}{if TRUE, uses the question label to label the column, if FALSE, displays the value labels}

\item{labelrowvars}{if TRUE, uses the question label to label the row, if FALSE, displays the value labels}

\item{weight}{weight variable used to create percent distribution, default is population weight (final_weight)}

\item{missing}{if TRUE, includes missing values in the denominator when calculating frequencies}

\item{breaklabels}{specifies a list of strings for labeling the breaks, where the first object in the string corresponds to the first break}

\item{overall}{if TRUE, includes overall row}

\item{roundn}{integer indicating the number of decimal places (round).}

\item{rowval}{for rowval [n], only display rows where the nth row has a value in numlist.}

\item{colval}{for colval [n], only displays columns where the nth column variable has a value in numlist}

\item{colname}{specifies the name of column [n], up to 5 columns}
}
\value{
the output will be a frequency table.
}
\description{
This function takes in a dataset and two lists of variables. It then reports the relative frequency
within its row of each cell. The rows are from row_vars and columns are from col_vars.
}
\examples{
shed_2019 <- read_sheddata(2019)
tabrow(shed_2019, row_vars = c(ppethm, educ_3cat), col_vars = c(EF1, EF2), rowvalall = c(1,2,3), colvalall = 1, labelcolvars = T)
}
