% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tabsubrow}
\alias{tabsubrow}
\title{Tabsubrow}
\usage{
#method for defualt
}
\arguments{
\item{shed_data}{SHED dataset.}

\item{subrow_var}{identifies SHED variable used for the subrow of the table}

\item{subrow_val}{only displays rows where the row variable has a value in numlist. Also sorts row display in the order of numbers in the numlist. default is to show all non-missing values.}

\item{subrow_labels}{optional argument to provide custom labels for the subrow breaks. Default is subrow_var labels.}

\item{tabsub_overall}{if TRUE includes overall row for the entire table}

\item{overall}{if TRUE, includes overall row within each group.}

\item{weight}{weight variable used to create percent distribution, default is population weight (final_weight)}

\item{row_vars}{list of SHED variables used for the rows of the table}

\item{col_vars}{list of SHED variables used fro the columns of the table}

\item{rowval[n]}{for rowval [n], only display rows where the nth row has a value in numlist; default is all values.}

\item{rowvalall}{for all row variables, only displays rows where each SHED variable in var_list has a value in rowvalall; deafult is all values.}

\item{colval[n]}{for colval [n], only display column where the nth column variable has a value in numlist; default is all values.}

\item{colvalall}{for all column variables, only displays columns where column variables have a value in colvalall;default is all values.}

\item{breaks}{inserts a row break before each new row}

\item{auto_labelbreaks}{if TRUE, displays row variables in row break lines. if FALSE, displays the value labels.}

\item{labelcolvars}{if TRUE, uses the question label to label the column, if FALSE, displays the value labels.}

\item{labelrowvars}{if TRUE, uses the question label to label the row, if FALSE, displays the value labels.}

\item{colname}{specifies the name of column [n], up to 5 columns}

\item{missing}{if TRUE, includes missing values in the denominator when calculating frequencies}

\item{breaklabels}{specifies a list of strings for labeling the breaks for rows, where the first object in the string corresponds to the first break}
}
\value{
the output will be a table
}
\description{
This function takes in a SHED dataset and two lists of variables. It then reports the relative frequency
}
\examples{
Replicates Table 3 from Chapter 5- Credit applicants in demographic groups by adverse credit outcome
table5.3<-tabsubrow(shed_2019, row_vars=inc_3cat, subrow_var = race_4cat, col_vars=c(A1_a,denied_offered_less), missing=FALSE,rowvalall=c(1,2,3),
subrow_val=c(1,2,3), colvalall=1, subtotals=TRUE)

}
