% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_excel}
\alias{write_excel}
\title{write_excel}
\usage{
write_excel(df, wb, excel, units = "Percent", title = "Table", reference = "sheet", footnote = "")
}
\arguments{
\item{df}{A data.frame or tibble object.}

\item{wb}{openxlsx::createWorkbook()}

\item{excel}{Excel file path. Requires a string.}

\item{units}{Requires a string.}

\item{title}{Title of table. Requires a string.}

\item{reference}{Reference type and number. Requires a string. Used for title and for sheetname.}

\item{footnote}{Table footnote. Requires a string.}

\item{title_ref}{Reference in the title if different from sheet reference. Requires a string.}
}
\value{
A workbook object
}
\description{
Writes a data.frame to an xlsx file.
}
\examples{
# Load in SHED 2019 data
SHED_2019<-data("shed2019")
# Create one-way table
rent_table <- multi1way(SHED_2019, c(R1_f, R1_e, R1_a, R1_c, R1_b, R1_d), weight = TRUE, missing = FALSE, labelvarnames = TRUE,
                 rowvalall = 1, breaks =c(1,3), breaklabels=c("Mortgage Access", "Preference"))
# Write table to excel
wb <- openxlsx::createWorkbook()
write_excel(rent_reason_table, wb, excel = "housing.xlsx", units = "Percent", title = "Reasons for Renting", reference = "Table 6.3", footnote = "Among renters. Respondents could select multiple answers.")
}
