% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{write_excel_plot}
\alias{write_excel_plot}
\title{write_excel_plot}
\usage{
write_excel_plot(g, wb, excel, reference, width = 6.5, height = 4)
}
\arguments{
\item{g}{A SHED plot/chart, or similar object.}

\item{wb}{openxlsx::createWorkbook()}

\item{excel}{Excel file path. Requires a string.}

\item{reference}{Reference type and number. Requires a string. Used for title and for sheetname.}
}
\value{
A workbook object
}
\description{
Writes a SHED plot/chart or ggplot object to an xlsx file.
}
\examples{
# Load in the SHED 2019 data
SHED_2019 <- data("shed2019")
# Creates a one-way table using multi1way()
rent_table <- multi1way(SHED_2019, c(R1_f, R1_e, R1_a, R1_c, R1_b, R1_d), weight = TRUE, missing = FALSE, labelvarnames = TRUE,
                 rowvalall = 1, breaks =c(1,3), breaklabels=c("Mortgage Access", "Preference"))
# Create bar chart from the one-way table
rent_chart <- oneway_chart(one_way_table, plot_title = "Reasons for Renting", plot_caption = "Among renters. Respondents could select multiple answers.")
# Write chart to excel
wb <- openxlsx::createWorkbook()
write_excel_plot(rent_chart, wb, excel = "housing.xlsx", reference = "Figure 23")
}
